[fix] move jwt to middleware (middleware.jwt.go)
